remove webauthn recommended concern, moved logic and tests
